+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.XQUARE.DayNight.NoActionBar"
         tools:targetApi="31">
         <activity
-            android:name="com.xquare.xquare_design_system_android.MainActivity"
+            android:name="com.xquare.xdsplayground.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.XQUARE.DayNight.NoActionBar">
