+++ xquare-design-system-android/src/main/kotlin/com/xquare/xdsandroid/typography/XdsTextView.kt
@@ -1,4 +1,4 @@
-package com.xquare.xds.typography
+package com.xquare.xdsandroid.typography
 
 import android.content.Context
 import android.content.res.TypedArray
@@ -7,7 +7,7 @@ import android.view.LayoutInflater
 import androidx.annotation.ColorRes
 import androidx.annotation.StyleRes
 import androidx.appcompat.widget.AppCompatTextView
-import com.xquare.xquare_design_system_android.R
+import com.xquare.xdsandroid.R
 
 @Deprecated(
     message = "Use 'style' attribute in layout",
@@ -19,8 +19,8 @@ public abstract class XdsTextView @JvmOverloads public constructor(
     @StyleRes private val style: Int,
     @ColorRes private val color: Int = R.color.neutral_0,
 ) : AppCompatTextView(
-    /* context = */ context,
-    /* attrs = */ attrs,
+    context,
+    attrs,
 ) {
 
     private var typedArray: TypedArray
@@ -28,12 +28,12 @@ public abstract class XdsTextView @JvmOverloads public constructor(
     init {
         val inflater = context.getSystemService(Context.LAYOUT_INFLATER_SERVICE) as LayoutInflater
         inflater.inflate(
-            /* resource = */ R.layout.typography,
-            /* root = */ findViewById(R.id.typography),
+            R.layout.typography,
+            findViewById(R.id.typography),
         )
         typedArray = context.obtainStyledAttributes(
-            /* set = */ attrs,
-            /* attrs = */ R.styleable.XdsTextView,
+            attrs,
+            R.styleable.XdsTextView,
         )
         setAttrs()
     }
@@ -56,8 +56,8 @@ public abstract class XdsTextView @JvmOverloads public constructor(
     private fun setTextColor() {
         setTextColor(
             typedArray.getColor(
-                /* index = */ R.styleable.XdsTextView_android_textColor,
-                /* defValue = */ resources.getColor(color),
+                R.styleable.XdsTextView_android_textColor,
+                resources.getColor(color),
             ),
         )
     }
