+++ xquare-design-system-android/src/main/kotlin/com/xquare/xdsandroid/typography/body/BodyLarge.kt
@@ -1,20 +1,20 @@
-package com.xquare.xds.typography.label
+package com.xquare.xdsandroid.typography.body
 
 import android.content.Context
 import android.util.AttributeSet
-import com.xquare.xds.typography.XdsTextView
-import com.xquare.xquare_design_system_android.R
+import com.xquare.xdsandroid.R
+import com.xquare.xdsandroid.typography.XdsTextView
 
 @Deprecated(
     message = "Use 'style' attribute in layout",
     level = DeprecationLevel.HIDDEN,
 )
-public class LabelMedium @JvmOverloads public constructor(
+public class BodyLarge @JvmOverloads public constructor(
     context: Context,
     attrs: AttributeSet? = null,
 ) : XdsTextView(
     context = context,
     attrs = attrs,
-    style = R.style.LabelMedium,
-    color = R.color.neutral_50,
+    style = R.style.BodyLarge,
+    color = R.color.neutral_30,
 )
