+++ xquare-design-system-android/src/main/kotlin/com/xquare/xdsandroid/typography/body/BodySmall.kt
@@ -1,20 +1,20 @@
-package com.xquare.xds.typography.title
+package com.xquare.xdsandroid.typography.body
 
 import android.content.Context
 import android.util.AttributeSet
-import com.xquare.xds.typography.XdsTextView
-import com.xquare.xquare_design_system_android.R
+import com.xquare.xdsandroid.R
+import com.xquare.xdsandroid.typography.XdsTextView
 
 @Deprecated(
     message = "Use 'style' attribute in layout",
     level = DeprecationLevel.HIDDEN,
 )
-public class TitleSmall @JvmOverloads public constructor(
+public class BodySmall @JvmOverloads public constructor(
     context: Context,
     attrs: AttributeSet? = null,
 ) : XdsTextView(
     context = context,
     attrs = attrs,
-    style = R.style.TitleSmall,
-    color = R.color.neutral_10,
+    style = R.style.BodySmall,
+    color = R.color.neutral_30,
 )
