+++ xquare-design-system-android/src/main/kotlin/com/xquare/xdsandroid/typography/display/DisplayMedium.kt
@@ -1,19 +1,19 @@
-package com.xquare.xds.typography.display
+package com.xquare.xdsandroid.typography.display
 
 import android.content.Context
 import android.util.AttributeSet
-import com.xquare.xds.typography.XdsTextView
-import com.xquare.xquare_design_system_android.R
+import com.xquare.xdsandroid.R
+import com.xquare.xdsandroid.typography.XdsTextView
 
 @Deprecated(
     message = "Use 'style' attribute in layout",
     level = DeprecationLevel.HIDDEN,
 )
-public class DisplaySmall @JvmOverloads public constructor(
+public class DisplayMedium @JvmOverloads public constructor(
     context: Context,
     attrs: AttributeSet? = null,
 ) : XdsTextView(
     context = context,
     attrs = attrs,
-    style = R.style.DisplaySmall,
+    style = R.style.DisplayMedium,
 )
