+++ xquare-design-system-android/src/main/kotlin/com/xquare/xdsandroid/typography/headline/HeadlineLarge.kt
@@ -1,9 +1,9 @@
-package com.xquare.xds.typography.headline
+package com.xquare.xdsandroid.typography.headline
 
 import android.content.Context
 import android.util.AttributeSet
-import com.xquare.xds.typography.XdsTextView
-import com.xquare.xquare_design_system_android.R
+import com.xquare.xdsandroid.R
+import com.xquare.xdsandroid.typography.XdsTextView
 
 @Deprecated(
     message = "Use 'style' attribute in layout",
