+++ xquare-design-system-android/src/main/kotlin/com/xquare/xdsandroid/typography/label/LabelLarge.kt
@@ -1,20 +1,20 @@
-package com.xquare.xds.typography.body
+package com.xquare.xdsandroid.typography.label
 
 import android.content.Context
 import android.util.AttributeSet
-import com.xquare.xds.typography.XdsTextView
-import com.xquare.xquare_design_system_android.R
+import com.xquare.xdsandroid.R
+import com.xquare.xdsandroid.typography.XdsTextView
 
 @Deprecated(
     message = "Use 'style' attribute in layout",
     level = DeprecationLevel.HIDDEN,
 )
-public class BodyLarge @JvmOverloads public constructor(
+public class LabelLarge @JvmOverloads public constructor(
     context: Context,
     attrs: AttributeSet? = null,
 ) : XdsTextView(
     context = context,
     attrs = attrs,
-    style = R.style.BodyLarge,
-    color = R.color.neutral_30,
+    style = R.style.LabelLarge,
+    color = R.color.neutral_50,
 )
